reorganize api routes
