+++ app/api/checks/[checkId]/[inspectorId]/info/route.js
