+++ app/api/checks/[checkId]/info/route.js
@@ -1,10 +1,10 @@
 import { createClient } from '@/app/supabase/service'
+import { createClient as createAdminClient } from '@/app/supabase/server'
 
 export async function GET(req, { params: { checkId } }) {
-  const userId = req.headers.get('x-user-id')
-  if (!userId) {
-    return Response.json({ error: 'Unauthorized' }, { status: 401 })
-  }
+  const adminSupabase = createAdminClient()
+  const res = await adminSupabase.auth.getUser()
+  const userId = res?.data?.user?.id ?? null
   const supabase = createClient()
   if (!checkId) {
     return Response.json({ error: 'Missing checkId parameter' }, { status: 400 })
