+++ app/api/projects/[projectId]/books/[bookId]/info/route.js
@@ -1,0 +1,87 @@
+import { createClient } from '@/app/supabase/service'
+
+/**
+ * @swagger
+ * /api/projects/{projectId}/books/{bookId}/info:
+ *   get:
+ *     summary: Get count of notes for book
+ *     tags:
+ *       - Books
+ *     parameters:
+ *       - in: path
+ *         name: projectId
+ *         required: true
+ *         schema:
+ *           type: integer
+ *       - in: path
+ *         name: bookId
+ *         required: true
+ *         schema:
+ *           type: integer
+ *     responses:
+ *       200:
+ *         description: Notes count
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 notes_count:
+ *                   type: integer
+ *                   description: Notes count for book
+ *       401:
+ *         description: Unauthorized
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 error:
+ *                   type: string
+ *                   description: Unauthorized
+ *       400:
+ *         description: Invalid request
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 error:
+ *                   type: string
+ *                   description: Invalid request
+ *       500:
+ *         description: Internal Server Error
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 error:
+ *                   type: string
+ *                   description: Internal Server Error
+ */
+
+export async function GET(req, { params: { bookId } }) {
+  const userId = req.headers.get('x-user-id')
+  if (!userId) {
+    return Response.json({ error: 'Unauthorized' }, { status: 401 })
+  }
+  const supabase = createClient()
+  if (!bookId) {
+    return Response.json({ error: 'Missing bookId parameter' }, { status: 400 })
+  }
+  try {
+    const { data, error } = await supabase.rpc('get_notes_count_for_book', {
+      book_id: bookId,
+      user_id: userId,
+    })
+
+    if (error) {
+      throw error
+    }
+
+    return Response.json(data, { status: 200 })
+  } catch (error) {
+    return Response.json({ error }, { status: 500 })
+  }
+}
