+++ middleware.js
@@ -11,7 +11,11 @@ acceptLanguage.languages(languages)
 
 export async function middleware(req) {
   const requestHeaders = new Headers(req.headers)
+
   let lng
+  // В этом блоке мы ищем язык запроса. Если в cookie есть язык,
+  // то берем его. Если нет, то берем язык из заголовка запроса.
+  // Если и в заголовке нет, то берем fallback-язык
   if (req.cookies.has(cookieName)) {
     lng = acceptLanguage.get(req.cookies.get(cookieName).value)
   }
@@ -22,11 +26,10 @@ export async function middleware(req) {
     lng = fallbackLng
   }
 
-  if (
-    // !['.png', '.jpg', '.ico', '.svg'].includes(req.nextUrl.pathname.slice(-4)) &&
-    !req.nextUrl.pathname.startsWith('/api') &&
-    !req.nextUrl.pathname.startsWith('/doc')
-  ) {
+  // этот код для всех страниц, кроме апи
+  // ниже в matcher мы исключили картинки, документацию, статичные файлы
+  if (!req.nextUrl.pathname.startsWith('/api')) {
+    // Если язык не указан в пути, то редиректим на страницу с языком
     if (
       !languages.some((loc) => req.nextUrl.pathname.startsWith(`/${loc}`)) &&
       !req.nextUrl.pathname.startsWith('/_next')
@@ -34,6 +37,7 @@ export async function middleware(req) {
       return NextResponse.redirect(new URL(`/${lng}${req.nextUrl.pathname}`, req.url))
     }
 
+    // Если в referer есть язык, то записываем его в cookie
     if (req.headers.has('referer')) {
       const refererUrl = new URL(req.headers.get('referer'))
       const lngInReferer = languages.find((l) => refererUrl.pathname.startsWith(`/${l}`))
@@ -45,6 +49,8 @@ export async function middleware(req) {
     }
   }
 
+  // для всех путей, которые ведут на страницу проекта или страницу токенов, но кроме апи.
+  // То есть это админская часть
   if (
     !req.nextUrl.pathname.startsWith('/api') &&
     (req.nextUrl.pathname.split('/')?.[2] === 'projects' ||
@@ -63,6 +69,7 @@ export async function middleware(req) {
         throw Error('Middleware error: Unauthorized user')
       }
     } catch (error) {
+      // Если пользователь не аутентифицирован, то редиректим на страницу логина
       const redirectUrl = req.nextUrl.clone()
       redirectUrl.pathname = '/' + lng + '/login'
       redirectUrl.searchParams.set(`redirectedFrom`, req.nextUrl.pathname)
@@ -71,15 +78,21 @@ export async function middleware(req) {
     return response
   }
 
-  // Доступ к API у нас в двух случаях. Или это аутентифицированный пользователь, или же в заголовке запроса есть токен. Проверим, и в зависимости от этого вернем ответ. Или ошибку доступа, или запрос, в который добавим айди пользователя.
+  // Доступ к API у нас в двух случаях. Или это аутентифицированный пользователь,
+  // или же в заголовке запроса есть токен. Проверим, и в зависимости от этого
+  // вернем ответ. Или ошибку доступа, или запрос, в который добавим айди пользователя.
   if (
     req.nextUrl.pathname.startsWith('/api') &&
-    !req.nextUrl.pathname.startsWith('/api/register')
+    !(
+      req.nextUrl.pathname.startsWith('/api/register') ||
+      req.nextUrl.pathname.startsWith('/api/checks')
+    )
   ) {
     let access_token
     try {
       access_token = req.headers.get('x-comcheck-token')
       if (!access_token) {
+        // Если токена нет, то проверяем аутентификацию пользователя
         const { supabase, response } = supabaseMiddleware(req)
         const { data, error } = await supabase.auth.getUser()
         if (error) {
@@ -102,11 +115,9 @@ export async function middleware(req) {
             cookies: new Cookies(response.cookies),
           },
         })
-        // response.headers.set('x-user-id', data.user?.id)
-
-        // return response
       }
     } catch (error) {
+      // Если аутентификация пользователя не удалась, то возвращаем ошибку
       return Response.json(
         { success: false, error, message: 'Password authentication failed' },
         { status: 401 }
@@ -132,6 +143,7 @@ export async function middleware(req) {
         },
       })
     } catch (error) {
+      // Если обработка токена не удалась, то возвращаем ошибку
       return Response.json(
         { success: false, message: 'Internal Server Error' },
         { status: 500 }
